+++ examples/complete/versions.tf
@@ -10,5 +10,9 @@ terraform {
       source  = "hashicorp/tls"
       version = ">= 3.2.0"
     }
+    archive = {
+      source  = "hashicorp/archive"
+      version = ">= 2.2.0"
+    }
   }
 }